refactor(gate): extract rule comparison owner
