+++ crates/tokmd-gate/src/evaluate.rs
@@ -1,8 +1,10 @@
 //! Policy evaluation logic.
 
-use crate::numeric::value_to_f64;
+mod compare;
+
 use crate::pointer::resolve_pointer;
 use crate::types::{GateResult, PolicyConfig, PolicyRule, RuleLevel, RuleOperator, RuleResult};
+use compare::{compare_contains, compare_equal, compare_in, compare_numeric};
 use serde_json::Value;
 
 /// Evaluate a policy against a JSON receipt.
@@ -165,76 +167,6 @@ fn evaluate_rule(receipt: &Value, rule: &PolicyRule, allow_missing: bool) -> Rul
         } else {
             rule.message.clone().or(generated_failure_message)
         },
-    }
-}
-
-/// Compare two values numerically.
-fn compare_numeric<F>(
-    actual: &Value,
-    expected: Option<&Value>,
-    cmp: F,
-) -> Result<bool, &'static str>
-where
-    F: Fn(f64, f64) -> bool,
-{
-    let actual_num = value_to_f64(actual).ok_or("actual value is not numeric")?;
-    let expected_num = expected
-        .and_then(value_to_f64)
-        .ok_or("expected value is missing or not numeric")?;
-    Ok(cmp(actual_num, expected_num))
-}
-
-/// Compare two values for equality.
-fn compare_equal(actual: &Value, expected: Option<&Value>) -> Result<bool, &'static str> {
-    let expected = expected.ok_or("expected value is missing")?;
-
-    // For strings, compare case-sensitively (before numeric to avoid "inf"/"nan" issues)
-    if let (Some(a), Some(b)) = (actual.as_str(), expected.as_str()) {
-        return Ok(a == b);
-    }
-
-    // For numeric types, compare as f64 to handle int/float mismatches
-    if let (Some(a), Some(b)) = (value_to_f64(actual), value_to_f64(expected)) {
-        return Ok((a - b).abs() < f64::EPSILON);
-    }
-
-    // For other types, use JSON equality
-    Ok(actual == expected)
-}
-
-/// Check if actual value is in a list of expected values.
-fn compare_in(actual: &Value, expected: Option<&Vec<Value>>) -> Result<bool, &'static str> {
-    let list = expected.ok_or("expected values list is missing")?;
-
-    for item in list {
-        if compare_equal(actual, Some(item)).unwrap_or(false) {
-            return Ok(true);
-        }
-    }
-
-    Ok(false)
-}
-
-/// Check if actual contains expected.
-fn compare_contains(actual: &Value, expected: Option<&Value>) -> Result<bool, &'static str> {
-    let expected = expected.ok_or("expected value is missing")?;
-
-    match actual {
-        Value::String(s) => {
-            let needle = expected
-                .as_str()
-                .ok_or("expected value must be a string for string contains checks")?;
-            Ok(s.contains(needle))
-        }
-        Value::Array(arr) => {
-            for item in arr {
-                if compare_equal(item, Some(expected)).unwrap_or(false) {
-                    return Ok(true);
-                }
-            }
-            Ok(false)
-        }
-        _ => Err("contains is only valid for string or array actual values"),
     }
 }
